+++ gridsome/package.json
@@ -87,6 +87,7 @@
     "slash": "^2.0.0",
     "sockjs": "^0.3.19",
     "sockjs-client": "^1.3.0",
+    "tapable": "^1.1.1",
     "url-loader": "^1.1.2",
     "vue": "^2.6.6",
     "vue-loader": "^15.6.0",
